experiments/pbn-research/requirements.txt
# PBN eval notebook + local pipelines (no paid APIs by default)
numpy>=1.24
scipy>=1.11
scikit-image>=0.22
scikit-learn>=1.3
pillow>=10.0
matplotlib>=3.8
pyyaml>=6.0
ipywidgets>=8.1
jupyter>=1.0.0
pandas>=2.0.0
pytest>=7.4.0
